requirements.txt
asgiref==3.3.1
astroid==2.4.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.3
django-ckeditor==6.0.0
django-debug-toolbar==3.1.1
django-js-asset==1.2.2
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.0.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
